+++ src/coleslaw.lisp
@@ -81,3 +81,12 @@ compile and deploy the blog."
     (compile-theme (theme *config*))
     (compile-blog (staging *config*))
     (deploy (staging *config*))))
+
+(defun preview (path &optional (content-type 'post))
+  "Render the content at PATH and save it to ~/tmp.html, loading the
+user's config and theme if necessary."
+  (unless *config*
+    (load-config nil)
+    (compile-theme (theme *config*)))
+  (let ((object (construct content-type (read-content path))))
+    (write-page "~/tmp.html" (render-page object))))