+++ NEWS.md
@@ -1,5 +1,6 @@
 ## Changes for 0.9.1 (2013-04-xx):
 
+* Added a PREVIEW function for REPL use.
 * Make ATOM and RSS templates a separate "generic" theme. (thanks @woudshoo!)
 
 ## Changes for 0.9 (2013-02-20):